requirements.txt
django=3.2.13
djangorestframework=3.13.1
